+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - View common email envelope headers with `:envelope`.
 - Notmuch accounts now support maildir operations: `:copy`, `:move`, `:mkdir`,
   `:rmdir`, `:archive` and the `copy-to` option.
+- Display messages from bottom to top with `reverse-msglist-order=true` in
+  `aerc.conf`.
 
 ### Fixed